quotes_labels: Restore previous utf-8 library
