+++ CMakeLists.txt
@@ -332,7 +332,6 @@ set(LIBNETDATA_FILES
         libnetdata/json/jsmn.h
         libnetdata/health/health.c
         libnetdata/health/health.h
-        libnetdata/string/utf8.c
         libnetdata/string/utf8.h
         libnetdata/socket/security.c
         libnetdata/socket/security.h)
