+++ libnetdata/Makefile.am
@@ -21,7 +21,6 @@ SUBDIRS = \
     socket \
     statistical \
     storage_number \
-    string \
     threads \
     url \
     tests \
